% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtscr_self_rank}
\alias{mtscr_self_rank}
\title{Self-chosen best answers}
\format{
\subsection{\code{mtscr_self_rank}}{

A tibble with 3225 rows and 4 columns:
\describe{
\item{subject}{patricipant's unique identification number}
\item{task}{divergent thinking task number}
\item{avr}{average judges' raiting}
\item{top_two}{indicator of self-chosen two best answer; 1 if chosen, 0 if not}
}
}
}
\source{
\url{https://osf.io/7rgsp/}
}
\usage{
mtscr_self_rank
}
\description{
An example dataset with best answers self-chosen by the participant. Use with \code{self_ranking}
argument in \link[=mtscr_model]{mtscr_model}.
}
\references{
\doi{10.1037/aca0000571}
}
\keyword{datasets}
