% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtscr-package.R
\docType{package}
\name{mtscr-package}
\alias{mtscr}
\alias{mtscr-package}
\title{mtscr: Multidimensional Top Scoring for Creativity Research}
\description{
Implementation of Multidimensional Top Scoring method for creativity assessment proposed in Boris Forthmann, Maciej Karwowski, Roger E. Beaty (2023) \doi{10.1037/aca0000571}.
}
\author{
\strong{Maintainer}: Jakub Jędrusiak \email{kuba23031999@gmail.com} (\href{https://orcid.org/0000-0002-6481-8210}{ORCID}) (University of Wrocław) [copyright holder]

Authors:
\itemize{
  \item Boris Forthmann \email{boris.forthmann@uni-muenster.de} (\href{https://orcid.org/0000-0001-9755-7304}{ORCID}) (University of Münster) [reviewer]
  \item Roger E. Beaty \email{rebeaty@psu.edu} (\href{https://orcid.org/0000-0001-6114-5973}{ORCID}) (Pennsylvania State University)
  \item Maciej Karwowski \email{maciej.karwowski@uwr.edu.pl} (\href{https://orcid.org/0000-0001-6974-1673}{ORCID}) (University of Wrocław)
}

}
\keyword{internal}
