% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtscr_creativity}
\alias{mtscr_creativity}
\title{Creativity assessment through semantic distance dataset}
\format{
## `mtscr_creativity`
A `tibble` with 4585 rows and 10 columns:
\describe{
  \item{id}{patricipants' unique identification number}
  \item{response}{response in AUT}
  \item{item}{item for which alternative uses were searched for}
  \item{SemDis_cbowukwacsubtitle_nf_m}{}
  \item{SemDis_cbowsubtitle_nf_m}{}
  \item{SemDis_cbowBNCwikiukwac_nf_m}{}
  \item{SemDis_TASA_nf_m}{}
  \item{SemDis_glove_nf_m}{}
  \item{SemDis_MEAN}{}
  \item{SemDis_factor}{}
}
}
\source{
<https://osf.io/7rgsp/>
}
\usage{
mtscr_creativity
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
A dataset from Forthmann, Karwowski & Beaty ([2023](https://doi.org/10.1037/aca0000571)) paper.
It contains a set of responses in Alternative Uses Task for different items with their
semantic distance assessment.
}
\keyword{datasets}
