% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtscr_prepare.r
\name{mtscr_prepare}
\alias{mtscr_prepare}
\title{Prepare database for MTS}
\usage{
mtscr_prepare(
  df,
  .id_column,
  .item_column,
  .value_column,
  preserve_existing = TRUE
)
}
\arguments{
\item{df}{Data frame in long format}

\item{.id_column}{Name of the column containing participants' id}

\item{.item_column}{Name of the column containing distinct trials (e.g. names of items in AUT)}

\item{.value_column}{Name of the column containing divergent thinking scores (e.g. semantic distance)}

\item{preserve_existing}{Logical, return the whole database with additional columns (`TRUE`) or only the additional columns (`FALSE`)}
}
\value{
The input data frame with additional columns:
    \describe{
        \item{`max_ind`}{Numerical, 0 for the best answer}
        \item{`top2_ind`}{Numerical, 0 for the two best answers}
        \item{`rank`}{ranking of the answer relative to participant and item}
    }
    The values are relative to the participant AND item, so the values for different participants scored for different tasks (e.g. uses for "brick" and "can") are distinct.
}
\description{
Prepare database for MTS
}
\examples{
data("mtscr_creativity", package = "mtscr")
mtscr_prepare(mtscr_creativity, id, item, SemDis_MEAN) # columns' names can also be "quoted"
}
