% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtscr_prepare.R
\name{mtscr_prepare}
\alias{mtscr_prepare}
\title{Prepare database for MTS}
\usage{
mtscr_prepare(df, id_column, item_column, score_column, minimal = FALSE)
}
\arguments{
\item{df}{Data frame in long format}

\item{id_column}{Name of the column containing participants' id}

\item{item_column}{Name of the column containing distinct trials (e.g. names of items in AUT)}

\item{score_column}{Name of the column containing divergent thinking scores (e.g. semantic distance)}

\item{minimal}{Logical, append columns to df (\code{FALSE}) or return only \code{id}, \code{item}, and the new columns (\code{TRUE})}
}
\value{
The input data frame with additional columns:
\describe{
\item{\code{.data$.z_score}}{Numerical, z-score of the creativity score}
\item{\code{.data$.ordering}}{Numerical, ranking of the answer relative to participant and item}
\item{\code{.max_ind}}{Numerical, 0 for the best answer, 1 otherwise}
\item{\code{.top2_ind}}{Numerical, 0 for the two best answers, 1 otherwise}
}
The values are relative to the participant AND item, so the values for different participants
scored for different tasks (e.g. uses for "brick" and "can") are distinct.
}
\description{
Prepare database for MTS analysis.
}
\examples{
data("mtscr_creativity", package = "mtscr")
mtscr_prepare(mtscr_creativity, id, item, SemDis_MEAN, minimal = TRUE)
}
