% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtscr_score.R
\name{mtscr_score}
\alias{mtscr_score}
\title{Score creativity with MTS}
\usage{
mtscr_score(
  df,
  id_column,
  item_column,
  score_column,
  model_type = c("all_max", "all_top2")
)
}
\arguments{
\item{df}{Data frame in long format}

\item{id_column}{Name of the column containing participants' id}

\item{item_column}{Name of the column containing distinct trials (e.g. names of items in AUT)}

\item{score_column}{Name of the column containing divergent thinking scores (e.g. semantic distance)}

\item{model_type}{Character vector, the type of model to fit. Can be \code{all_max}, \code{all_top2} or both.}
}
\value{
A data frame with the creativity scores in the columns \code{.all_max} and \code{.all_top2}.
}
\description{
Score creativity with MTS
}
\examples{
data("mtscr_creativity", package = "mtscr")
mtscr_score(mtscr_creativity, id, item, SemDis_MEAN)
}
